refactor: rename nextFloat to nextDouble for consistency in input handling

+++ main.ts
@@ -4,7 +4,7 @@ const scanner = new Scanner();
 
 const nome = await scanner.nextLine("Qual seu nome? ");
 const idade = await scanner.nextInt("Qual sua idade? ");
-const altura = await scanner.nextFloat("Qual sua altura? ");
+const altura = await scanner.nextDouble("Qual sua altura? ");
 
 console.log(`Olá, ${nome}. Você tem ${idade} anos e mede ${altura}m.`);
